+++ test/unit/util.test.ts
@@ -2,7 +2,7 @@
  * this tests the behaviour of util.js
  */
 import assert from 'assert';
-import AsyncTestUtil from 'async-test-util';
+import AsyncTestUtil, { wait } from 'async-test-util';
 import {
     fastUnsecureHash,
     randomCouchString,
@@ -17,8 +17,11 @@ import {
     defaultHashFunction,
     b64DecodeUnicode,
     b64EncodeUnicode,
-    batchArray
+    batchArray,
+    clone as rxdbClone
 } from '../../';
+import config from './config';
+
 import {
     validateDatabaseName,
     deepFreezeWhenDevMode
@@ -146,6 +149,53 @@ describe('util.test.js', () => {
             };
             const sorted = sortObject(obj);
             assert.ok(sorted.color.$regex instanceof RegExp);
+        });
+    });
+    describe('.recursiveDeepCopy()', () => {
+        /**
+         * Test the performance of different methods.
+         */
+        const cloneMethods: ((o: any) => any)[] = [
+            o => rxdbClone(o),
+            o => structuredClone(o),
+            o => JSON.parse(JSON.stringify(o))
+        ];
+        cloneMethods.forEach(method => {
+            it('run once', async () => {
+                if(!config.isFastMode()){
+                    await wait(200);
+                }
+                let obj = {
+                    a: 'a',
+                    b: 7879,
+                    g: false,
+                    h: [
+                        { g: 0, l: true, k: { l: 56 } },
+                        { g: 0, l: true, fg: 'dfg', k: { l: 56 } },
+                        { g: 0, l: true, k: { l: 56 } }
+                    ],
+                    jk: {
+                        ager: 56,
+                        tank: 'sj',
+                        what: {
+                            sdf: {
+                                sdf: 'asdasd',
+                                ll: ['safd', { sdfs: 'dfsf' }]
+                            }
+                        }
+                    }
+                };
+                const start = performance.now();
+                let t = 0;
+                const runs = config.isFastMode() ? 100 : 2000;
+                while (t < runs) {
+                    t++;
+                    method(obj);
+                    obj = Object.assign({}, obj);
+                }
+                const time = performance.now() - start;
+                console.log('time ' + time);
+            });
         });
     });
     describe('.validateDatabaseName()', () => {